chore(): move issue tests to overlay-components/
